@@ -2,17 +2,19 @@
   "name": "resume",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "devDependencies": {
-    "react-scripts": "0.8.4"
+    "@vitejs/plugin-react": "^4.7.0",
+    "vite": "^7.0.5"
   },
   "dependencies": {
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1"
+    "markdown": "^0.5.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   }
 }
